sensors/dht22/requirements.txt
flask
py-healthcheck
psutil
